HSV calibration program
For tracking eyeball movement, we need to filter the pupil from the rest of the eye. In order to that we need to threshold HSV value of pupil from the the frame. These values are light dependent and cahnge with light. Hence for a particular lighting condition we need to get the right HSV values for optimal effieciency. Hence with this program we get the required HSV values.